feat(F-9,P2-24): run PDF operations in an inlined Web Worker
Every PDF operation ran on the main thread, so a large document froze the
tab with no feedback and no way to cancel (P2-24). They now run in a
worker.

The single-file distribution makes this less routine than usual. A worker
referenced as a sibling file could never be fetched from a null origin,
so it is inlined via ?worker&inline. It must also be a *classic* worker:
module workers were measured failing to start from file://, hence
worker.format 'iife' in the vite config.

That constraint had a consequence worth recording. IIFE output cannot
code-split, and html2pdf.js does, so pulling the DOCX path into the
worker's import graph broke the build outright. convertDocxToPdf now
lives in its own module -- it renders through html2canvas and needs a DOM
it would not have in a worker anyway, so it stays on the main thread by
design rather than by accident.

Structure: pdf-ops.ts holds the implementations, pdf.worker.ts dispatches
to them by name, and pdf-utils.ts is the public API that routes through
the worker and falls back to an inline call when one is unavailable.
Tests target pdf-utils and so exercise the fallback, since jsdom provides
no Worker; the worker path is covered by the browser checks.

The fallback distinguishes transport failures from operation failures. A
worker that cannot start must not take the feature down with it, but a
genuine error -- a wrong password, an invalid page range -- must surface
as itself rather than being silently re-run on the main thread.

Cost: the bundle grows 2.64MB to 3.52MB, because the worker carries its
own copy of pdf-lib and cannot share the main chunk's.

Verified against the file:// build: a real tool operation spawns a
blob-URL worker, completes, and downloads, while the main thread stays
fully responsive (60 timer ticks in 600ms). 155 tests pass, typecheck
clean, lint unchanged, zero network requests.
